refactoring
se agrega toda la funcionalidad de la simulación en el directorio /des
