+++ des/singleton.py
